test/K8sServices: Tests for UDP connectivity
Adds an additional UDP echoserver to the pods and service definitions,
allowing us to check UDP connectivity alongside the existing TCP tests.

The new echoserver uses the TFTP protocol to serve a status page similar
to the existing HTTP server used for TCP tests. This allows us to reuse
the existing test infrastructure, i.e. curl has built-in support for
TFTP.

The deployed cilium/echoserver-udp is using the so-called "single-port"
mode of TFTP, where the server will always answer on the UDP port on
which it received the request. This is required for NAT to work.

Fixes: #9363

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>